requirements.txt
Django==5.0.6
djangorestframework==3.15.1
Flask==3.0.3
requests==2.31.0
boto3==1.34.0